Made slight changes to everything. (See desc) The most notable thing is that the Sandbox project was deleted as it was just sitting there, unused.
TODO:
1) Joystick Input
2) Decide if tapping into Monitor callbacks is
useful. If not, remove it and revert the change
I made to ImGui's monitor.c (ctrl+f
g_PrevUserMonitorCallback or something like that)

+++ GBC/src/Platform/Renderer/OpenGL/OpenGLVertexArray.cpp
@@ -34,7 +34,7 @@ namespace gbc
 	
 		const BufferLayout& layout = vertexBuffer->GetLayout();
 		const std::vector<BufferElement>& elements = layout.GetElements();
-		for (unsigned int i = 0; i < elements.size(); i++)
+		for (uint32_t i = 0; i < static_cast<uint32_t>(elements.size()); i++)
 		{
 			const BufferElement& element = elements[i];
 			glEnableVertexAttribArray(i);
